requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
psycopg[binary]>=3.1,<4.0
Pillow>=10.0,<12.0
python-dotenv>=1.0,<2.0